cleanup first version of tutorials
