+++ docs/source/auto_tutorials/sg_execution_times.rst
@@ -5,12 +5,12 @@
 
 Computation times
 =================
-**02:18.109** total execution time for **auto_tutorials** files:
+**02:17.939** total execution time for **auto_tutorials** files:
 
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py` (``tutorial_image_based_optimization_with_pyramid.py``) | 01:24.500 | 0.0 MB |
+| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization_with_pyramid.py` (``tutorial_image_based_optimization_with_pyramid.py``) | 01:24.973 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization.py` (``tutorial_image_based_optimization.py``)                           | 00:53.609 | 0.0 MB |
+| :ref:`sphx_glr_auto_tutorials_tutorial_image_based_optimization.py` (``tutorial_image_based_optimization.py``)                           | 00:52.966 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_auto_tutorials_tutorial_nst_without_pystiche.py` (``tutorial_nst_without_pystiche.py``)                                   | 00:00.000 | 0.0 MB |
 +------------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
