+++ docs/source/auto_tutorials/tutorial_image_based_optimization.py
@@ -14,7 +14,7 @@ from pystiche.image import show_image, write_image
 from pystiche.enc import vgg19_encoder
 from pystiche.ops import MSEEncodingOperator, GramOperator, MultiLayerEncodingOperator
 from pystiche.loss import MultiOperatorLoss
-from utils import demo_images
+from pystiche.demo import demo_images
 
 
 ###############################################################################
