+++ docs/source/auto_tutorials/tutorial_image_based_optimization_with_pyramid.py.md5
@@ -1,1 +1,1 @@
-4c04d05ba585153e1ce591ccd4541ac7
+9fc3939c2f7bfa4cfbebd77bf42e8ba5
